+++ test/shell.js
@@ -1,37 +1,37 @@
-import shell from '../dist/shell-escape-tag.js'
+import shell from '..'
 import test  from 'ava'
 
 test('escapes an empty string', t => {
-    t.true(shell`` === '')
+    t.is(shell``, '')
 })
 
 test('ignores null values', t => {
-    t.true(shell`foo${null}bar` === 'foobar')
+    t.is(shell`foo${null}bar`, 'foobar')
 })
 
 test('ignores undefined values', t => {
-    t.true(shell`foo${undefined}bar` === 'foobar')
+    t.is(shell`foo${undefined}bar`, 'foobar')
 })
 
 test('ignores null and undefined values', t => {
     const bar = [ null, undefined, 'bar', undefined, null ]
-    t.true(shell`foo${bar}baz` === 'foobarbaz')
+    t.is(shell`foo${bar}baz`, 'foobarbaz')
 })
 
 test('does not ignore empty strings', t => {
     const bar = [ '', 'bar', '' ]
-    t.true(shell`foo${bar}baz` === 'foo bar baz')
+    t.is(shell`foo${bar}baz`, 'foo bar baz')
 })
 
 test('escapes a string with no interpolations', t => {
-    t.true(shell`foo` === 'foo')
-    t.true(shell`foo bar` === 'foo bar')
-    t.true(shell`foo bar baz` === 'foo bar baz')
+    t.is(shell`foo`, 'foo')
+    t.is(shell`foo bar`, 'foo bar')
+    t.is(shell`foo bar baz`, 'foo bar baz')
 })
 
 test('escapes a string which only contains an interpolation', t => {
     const foo = 'Foo'
-    t.true(shell`${foo}` === 'Foo')
+    t.is(shell`${foo}`, 'Foo')
 })
 
 test('escapes a string which only contains interpolations', t => {
@@ -39,14 +39,14 @@ test('escapes a string which only contains interpolations', t => {
     const bar = 'Bar'
     const baz = 'Baz'
 
-    t.true(shell`${foo}${bar}${baz}` === 'FooBarBaz')
+    t.is(shell`${foo}${bar}${baz}`, 'FooBarBaz')
 })
 
 test('escapes a string which starts with an interpolation', t => {
     const foo = 'Foo'
 
-    t.true(shell`${foo}bar` === 'Foobar')
-    t.true(shell`${foo} bar` === 'Foo bar')
+    t.is(shell`${foo}bar`, 'Foobar')
+    t.is(shell`${foo} bar`, 'Foo bar')
 })
 
 test('escapes a string which starts with interpolations', t => {
@@ -54,15 +54,15 @@ test('escapes a string which starts with interpolations', t => {
     const bar  = 'Bar'
     const baz  = 'Baz'
 
-    t.true(shell`${foo}${bar}${baz}quux` === 'FooBarBazquux')
-    t.true(shell`${foo} ${bar} ${baz} quux` === 'Foo Bar Baz quux')
+    t.is(shell`${foo}${bar}${baz}quux`, 'FooBarBazquux')
+    t.is(shell`${foo} ${bar} ${baz} quux`, 'Foo Bar Baz quux')
 })
 
 test('escapes a string which ends with an interpolation', t => {
     const foo = 'Foo'
 
-    t.true(shell`foo${foo}` === 'fooFoo')
-    t.true(shell`foo ${foo}` === 'foo Foo')
+    t.is(shell`foo${foo}`, 'fooFoo')
+    t.is(shell`foo ${foo}`, 'foo Foo')
 })
 
 test('escapes a string which ends with interpolations', t => {
@@ -70,18 +70,18 @@ test('escapes a string which ends with interpolations', t => {
     const bar = 'Bar'
     const baz = 'Baz'
 
-    t.true(shell`foo${foo}${bar}${baz}` === 'fooFooBarBaz')
-    t.true(shell`foo ${foo} ${bar} ${baz}` === 'foo Foo Bar Baz')
+    t.is(shell`foo${foo}${bar}${baz}`, 'fooFooBarBaz')
+    t.is(shell`foo ${foo} ${bar} ${baz}`, 'foo Foo Bar Baz')
 })
 
 test('escapes a string with spaces', t => {
     const foo = 'Foo Bar'
-    t.true(shell`foo ${foo}` === "foo 'Foo Bar'")
+    t.is(shell`foo ${foo}`, "foo 'Foo Bar'")
 })
 
 test('escapes an array of strings with spaces', t => {
     const foo = [ 'Foo Bar', 'Baz Quux' ]
-    t.true(shell`foo ${foo}` === "foo 'Foo Bar' 'Baz Quux'")
+    t.is(shell`foo ${foo}`, "foo 'Foo Bar' 'Baz Quux'")
 })
 
 test('escapes a string with quotes', t => {